refactor: split host route helpers
