+++ src/backend/database/routes/host.ts
@@ -38,6 +38,18 @@ import { SystemCrypto } from "../../utils/system-crypto.js";
 import { DatabaseSaveTrigger } from "../db/index.js";
 import { parseSSHKey } from "../../utils/ssh-key-utils.js";
 import { sendWakeOnLan, isValidMac } from "../../utils/wake-on-lan.js";
+import {
+  isNonEmptyString,
+  isValidPort,
+  normalizeImportedHost,
+  stripSensitiveFields,
+  transformHostResponse,
+} from "./host-normalizers.js";
+import {
+  normalizeSelectOpParam,
+  renderOpksshErrorPage,
+  rewriteOPKSSHHtml,
+} from "./opkssh-html.js";
 
 const router = express.Router();
 
@@ -69,279 +81,6 @@ function notifyStatsHostUpdated(
         error: err instanceof Error ? err.message : String(err),
       });
     });
-}
-
-function isNonEmptyString(value: unknown): value is string {
-  return typeof value === "string" && value.trim().length > 0;
-}
-
-function isValidPort(port: unknown): port is number {
-  return typeof port === "number" && port > 0 && port <= 65535;
-}
-
-function asString(value: unknown): string | undefined {
-  return typeof value === "string" && value.trim() ? value.trim() : undefined;
-}
-
-function asPort(value: unknown): number | undefined {
-  const port =
-    typeof value === "number"
-      ? value
-      : typeof value === "string"
-        ? Number.parseInt(value, 10)
-        : NaN;
-
-  return isValidPort(port) ? port : undefined;
-}
-
-function asInteger(value: unknown): number | undefined {
-  const number =
-    typeof value === "number"
-      ? value
-      : typeof value === "string"
-        ? Number.parseInt(value, 10)
-        : NaN;
-
-  return Number.isInteger(number) ? number : undefined;
-}
-
-function asBoolean(value: unknown, fallback = false): boolean {
-  if (typeof value === "boolean") return value;
-  if (typeof value === "number") return value !== 0;
-  if (typeof value === "string") {
-    const normalized = value.trim().toLowerCase();
-    if (["true", "1", "yes", "on"].includes(normalized)) return true;
-    if (["false", "0", "no", "off"].includes(normalized)) return false;
-  }
-
-  return fallback;
-}
-
-function normalizeImportTags(value: unknown): string[] {
-  if (Array.isArray(value)) {
-    return value
-      .map((tag) => asString(tag))
-      .filter((tag): tag is string => !!tag);
-  }
-  if (typeof value === "string") {
-    return value
-      .split(",")
-      .map((tag) => tag.trim())
-      .filter(Boolean);
-  }
-
-  return [];
-}
-
-type NormalizedImportedHost = Record<string, unknown> & {
-  connectionType: string;
-  name?: string;
-  ip?: string;
-  port: number;
-  username?: string;
-  folder?: string;
-  tags: string[];
-  authType?: string;
-  password?: string;
-  key?: string;
-  keyPassword?: string;
-  keyType?: string;
-  credentialId?: number;
-  pin?: unknown;
-  enableTerminal?: unknown;
-  enableTunnel?: unknown;
-  enableFileManager?: unknown;
-  enableDocker?: unknown;
-  showTerminalInSidebar?: unknown;
-  showFileManagerInSidebar?: unknown;
-  showTunnelInSidebar?: unknown;
-  showDockerInSidebar?: unknown;
-  showServerStatsInSidebar?: unknown;
-  defaultPath?: unknown;
-  sudoPassword?: unknown;
-  tunnelConnections?: unknown;
-  jumpHosts?: unknown;
-  quickActions?: unknown;
-  statsConfig?: unknown;
-  dockerConfig?: unknown;
-  terminalConfig?: unknown;
-  forceKeyboardInteractive?: unknown;
-  notes?: unknown;
-  useSocks5?: unknown;
-  socks5Host?: unknown;
-  socks5Port?: unknown;
-  socks5Username?: unknown;
-  socks5Password?: unknown;
-  socks5ProxyChain?: unknown;
-  portKnockSequence?: unknown;
-  overrideCredentialUsername?: unknown;
-  domain?: unknown;
-  security?: unknown;
-  ignoreCert?: unknown;
-  guacamoleConfig?: unknown;
-  enableSsh: boolean;
-  enableRdp: boolean;
-  enableVnc: boolean;
-  enableTelnet: boolean;
-};
-
-function normalizeImportedHost(
-  hostData: Record<string, unknown>,
-): NormalizedImportedHost {
-  const connectionType =
-    asString(hostData.connectionType) ||
-    (asBoolean(hostData.enableRdp)
-      ? "rdp"
-      : asBoolean(hostData.enableVnc)
-        ? "vnc"
-        : asBoolean(hostData.enableTelnet)
-          ? "telnet"
-          : "ssh");
-
-  const port =
-    asPort(hostData.port) ||
-    (connectionType === "rdp"
-      ? asPort(hostData.rdpPort) || 3389
-      : connectionType === "vnc"
-        ? asPort(hostData.vncPort) || 5900
-        : connectionType === "telnet"
-          ? asPort(hostData.telnetPort) || 23
-          : asPort(hostData.sshPort) || 22);
-
-  return {
-    ...hostData,
-    connectionType,
-    name: asString(hostData.name) || asString(hostData.label),
-    ip:
-      asString(hostData.ip) ||
-      asString(hostData.address) ||
-      asString(hostData.host) ||
-      asString(hostData.hostname),
-    port,
-    username: asString(hostData.username) || asString(hostData.user),
-    folder: asString(hostData.folder) || asString(hostData.group),
-    tags: normalizeImportTags(hostData.tags),
-    credentialId: asInteger(hostData.credentialId),
-    authType:
-      asString(hostData.authType) ||
-      asString(hostData.authMethod) ||
-      (hostData.credentialId ? "credential" : hostData.key ? "key" : undefined),
-    enableSsh:
-      hostData.enableSsh === undefined
-        ? connectionType === "ssh"
-        : asBoolean(hostData.enableSsh),
-    enableRdp:
-      hostData.enableRdp === undefined
-        ? connectionType === "rdp"
-        : asBoolean(hostData.enableRdp),
-    enableVnc:
-      hostData.enableVnc === undefined
-        ? connectionType === "vnc"
-        : asBoolean(hostData.enableVnc),
-    enableTelnet:
-      hostData.enableTelnet === undefined
-        ? connectionType === "telnet"
-        : asBoolean(hostData.enableTelnet),
-  };
-}
-
-const SENSITIVE_FIELDS = [
-  "key",
-  "keyPassword",
-  "autostartKey",
-  "autostartKeyPassword",
-];
-
-function stripSensitiveFields(
-  host: Record<string, unknown>,
-): Record<string, unknown> {
-  const result = { ...host };
-  result.hasKey = !!host.key;
-  result.hasKeyPassword = !!host.keyPassword;
-  for (const field of SENSITIVE_FIELDS) {
-    delete result[field];
-  }
-  return result;
-}
-
-function transformHostResponse(
-  host: Record<string, unknown>,
-): Record<string, unknown> {
-  return {
-    ...host,
-    tags:
-      typeof host.tags === "string"
-        ? host.tags
-          ? host.tags.split(",").filter(Boolean)
-          : []
-        : [],
-    pin: !!host.pin,
-    enableTerminal: !!host.enableTerminal,
-    enableTunnel: !!host.enableTunnel,
-    enableFileManager: !!host.enableFileManager,
-    enableDocker: !!host.enableDocker,
-    showTerminalInSidebar: !!host.showTerminalInSidebar,
-    showFileManagerInSidebar: !!host.showFileManagerInSidebar,
-    showTunnelInSidebar: !!host.showTunnelInSidebar,
-    showDockerInSidebar: !!host.showDockerInSidebar,
-    showServerStatsInSidebar: !!host.showServerStatsInSidebar,
-    // Old hosts only had connection_type set; the per-protocol enable flags didn't exist yet.
-    // The schema defaults (enableSsh=true, others=false) wrongly mark every old host as SSH.
-    // Detect this migration case: if no non-SSH protocol is explicitly enabled AND
-    // connectionType is set to a non-SSH value, fall back to inferring from connectionType.
-    ...(() => {
-      const ct = host.connectionType;
-      const rdp = !!host.enableRdp;
-      const vnc = !!host.enableVnc;
-      const tel = !!host.enableTelnet;
-      const isMigratedNonSsh = !rdp && !vnc && !tel && ct && ct !== "ssh";
-      return {
-        enableSsh: isMigratedNonSsh ? false : !!host.enableSsh,
-        enableRdp: isMigratedNonSsh ? ct === "rdp" : rdp,
-        enableVnc: isMigratedNonSsh ? ct === "vnc" : vnc,
-        enableTelnet: isMigratedNonSsh ? ct === "telnet" : tel,
-      };
-    })(),
-    sshPort: host.sshPort ?? host.port ?? 22,
-    rdpPort: host.rdpPort ?? 3389,
-    vncPort: host.vncPort ?? 5900,
-    telnetPort: host.telnetPort ?? 23,
-    rdpUser: host.rdpUser || undefined,
-    rdpDomain: host.rdpDomain || undefined,
-    rdpSecurity: host.rdpSecurity || undefined,
-    rdpIgnoreCert: !!host.rdpIgnoreCert,
-    vncUser: host.vncUser || undefined,
-    telnetUser: host.telnetUser || undefined,
-    tunnelConnections: host.tunnelConnections
-      ? JSON.parse(host.tunnelConnections as string)
-      : [],
-    jumpHosts: host.jumpHosts ? JSON.parse(host.jumpHosts as string) : [],
-    quickActions: host.quickActions
-      ? JSON.parse(host.quickActions as string)
-      : [],
-    statsConfig: host.statsConfig
-      ? JSON.parse(host.statsConfig as string)
-      : undefined,
-    terminalConfig: host.terminalConfig
-      ? JSON.parse(host.terminalConfig as string)
-      : undefined,
-    dockerConfig: host.dockerConfig
-      ? JSON.parse(host.dockerConfig as string)
-      : undefined,
-    forceKeyboardInteractive: host.forceKeyboardInteractive === "true",
-    socks5ProxyChain: host.socks5ProxyChain
-      ? JSON.parse(host.socks5ProxyChain as string)
-      : [],
-    portKnockSequence: host.portKnockSequence
-      ? JSON.parse(host.portKnockSequence as string)
-      : [],
-    domain: host.domain || undefined,
-    security: host.security || undefined,
-    ignoreCert: !!host.ignoreCert,
-    guacamoleConfig: host.guacamoleConfig
-      ? JSON.parse(host.guacamoleConfig as string)
-      : undefined,
-  };
 }
 
 const authManager = AuthManager.getInstance();
@@ -4499,285 +4238,6 @@ router.delete(
     }
   },
 );
-
-function escapeHtml(value: string): string {
-  return value
-    .replace(/&/g, "&amp;")
-    .replace(/</g, "&lt;")
-    .replace(/>/g, "&gt;")
-    .replace(/"/g, "&quot;")
-    .replace(/'/g, "&#39;");
-}
-
-// Replicates openpubkey's client/choosers/web_chooser.go IssuerToName().
-// OPKSSH's /select handler keys its providerMap by this derived name, NOT by the
-// `alias` field in config.yml. We need the same mapping so we can normalize any
-// `op=` query param we receive (which can be alias, issuer with protocol, or
-// issuer without protocol depending on client version) to what OPKSSH expects.
-function opksshIssuerToName(issuer: string): string | null {
-  if (!issuer) return null;
-  const withScheme =
-    issuer.startsWith("http://") || issuer.startsWith("https://")
-      ? issuer
-      : `https://${issuer}`;
-  if (withScheme.startsWith("https://accounts.google.com")) return "google";
-  if (withScheme.startsWith("https://login.microsoftonline.com"))
-    return "azure";
-  if (withScheme.startsWith("https://gitlab.com")) return "gitlab";
-  if (withScheme.startsWith("https://issuer.hello.coop")) return "hello";
-  if (withScheme.startsWith("https://")) {
-    const host = withScheme.slice("https://".length).split("/")[0];
-    return host || null;
-  }
-  return null;
-}
-
-function normalizeSelectOpParam(
-  rawOp: string,
-  providers: Array<{ alias: string; issuer: string }>,
-): string {
-  if (!rawOp) return rawOp;
-  const knownNames = new Set(
-    providers
-      .map((p) => opksshIssuerToName(p.issuer))
-      .filter((n): n is string => typeof n === "string" && n.length > 0),
-  );
-  if (knownNames.has(rawOp)) return rawOp;
-
-  const derivedFromRaw = opksshIssuerToName(rawOp);
-  if (derivedFromRaw && knownNames.has(derivedFromRaw)) return derivedFromRaw;
-
-  const matchByAlias = providers.find((p) => p.alias === rawOp);
-  if (matchByAlias) {
-    const name = opksshIssuerToName(matchByAlias.issuer);
-    if (name) return name;
-  }
-
-  return rawOp;
-}
-
-interface OpksshErrorPageOptions {
-  title: string;
-  heading: string;
-  message: string;
-  details?: string;
-  requestId?: string;
-  statusCode?: number;
-}
-
-function renderOpksshErrorPage(opts: OpksshErrorPageOptions): string {
-  const title = escapeHtml(opts.title);
-  const heading = escapeHtml(opts.heading);
-  const message = escapeHtml(opts.message);
-  const detailsBlock = opts.details
-    ? `<pre class="details">${escapeHtml(opts.details)}</pre>`
-    : "";
-  const requestIdBlock = opts.requestId
-    ? `<p class="request-id">Request ID: ${escapeHtml(opts.requestId)}</p>`
-    : "";
-
-  return `<!DOCTYPE html>
-<html>
-<head>
-  <title>${title}</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <style>
-    * { margin: 0; padding: 0; box-sizing: border-box; }
-    body {
-      font-family: system-ui, -apple-system, BlinkMacSystemFont, 'Segoe UI', sans-serif;
-      display: flex;
-      justify-content: center;
-      align-items: center;
-      min-height: 100vh;
-      background: #18181b;
-      color: #fafafa;
-      padding: 1rem;
-    }
-    .container {
-      text-align: center;
-      background: #27272a;
-      padding: 3rem 2rem;
-      border-radius: 0.625rem;
-      border: 1px solid rgba(255, 255, 255, 0.1);
-      max-width: 720px;
-      width: 100%;
-    }
-    h1 {
-      color: #fafafa;
-      font-size: 1.5rem;
-      font-weight: 600;
-      margin-bottom: 0.75rem;
-    }
-    p {
-      color: #9ca3af;
-      font-size: 0.95rem;
-      line-height: 1.5;
-      white-space: pre-wrap;
-      word-break: break-word;
-    }
-    p + p { margin-top: 0.5rem; }
-    .details {
-      color: #d4d4d8;
-      text-align: left;
-      font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
-      font-size: 0.8rem;
-      line-height: 1.45;
-      margin-top: 1.25rem;
-      padding: 0.875rem 1rem;
-      background: #0f0f11;
-      border: 1px solid rgba(255, 255, 255, 0.08);
-      border-radius: 0.5rem;
-      white-space: pre-wrap;
-      word-break: break-word;
-      overflow-x: auto;
-    }
-    .request-id {
-      color: #6b7280;
-      font-size: 0.75rem;
-      margin-top: 1rem;
-      font-family: ui-monospace, SFMono-Regular, Menlo, Consolas, monospace;
-    }
-  </style>
-</head>
-<body>
-  <div class="container">
-    <h1>${heading}</h1>
-    <p>${message}</p>
-    ${detailsBlock}
-    ${requestIdBlock}
-  </div>
-</body>
-</html>`;
-}
-
-function rewriteOPKSSHHtml(
-  html: string,
-  requestId: string,
-  routePrefix: "opkssh-chooser" | "opkssh-callback",
-): string {
-  const basePath = `/host/${routePrefix}/${requestId}`;
-  const localHostPattern = "(?:localhost|127\\.0\\.0\\.1)";
-
-  const attrPatterns = ["action", "href", "src", "formaction"];
-  for (const attr of attrPatterns) {
-    html = html.replace(
-      new RegExp(`${attr}="(/[^"]*)`, "g"),
-      `${attr}="${basePath}$1`,
-    );
-    html = html.replace(
-      new RegExp(`${attr}='(/[^']*)`, "g"),
-      `${attr}='${basePath}$1`,
-    );
-  }
-
-  for (const attr of ["href", "action", "src", "formaction"]) {
-    html = html.replace(
-      new RegExp(
-        `${attr}=["']?http:\\/\\/${localHostPattern}:\\d+\\/([^"'\\s]*)`,
-        "g",
-      ),
-      `${attr}="${basePath}/$1`,
-    );
-  }
-
-  html = html.replace(
-    new RegExp(
-      `(window\\.location\\.href\\s*=\\s*["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["'])`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-  html = html.replace(
-    new RegExp(
-      `(window\\.location\\s*=\\s*["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["'])`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-  html = html.replace(
-    new RegExp(
-      `(fetch\\(["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["'])`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-
-  html = html.replace(
-    new RegExp(
-      `(location\\.assign\\(["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["']\\))`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-  html = html.replace(
-    new RegExp(
-      `(location\\.replace\\(["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["']\\))`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-
-  // XMLHttpRequest.open("GET", "http://localhost:PORT/path", ...)
-  html = html.replace(
-    new RegExp(
-      `(\\.open\\(["']\\w+["']\\s*,\\s*["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["'])`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-
-  html = html.replace(
-    new RegExp(
-      `(<meta[^>]+http-equiv=["']refresh["'][^>]+content=["'][^;]+;\\s*url=)http:\\/\\/${localHostPattern}:\\d+\\/([^"']+)(["'][^>]*>)`,
-      "gi",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-
-  html = html.replace(
-    new RegExp(
-      `(data-[\\w-]+=["'])http:\\/\\/${localHostPattern}:\\d+\\/([^"']*)(["'])`,
-      "g",
-    ),
-    `$1${basePath}/$2$3`,
-  );
-
-  const baseTag = `<base href="${basePath}/">`;
-
-  if (html.includes("<base")) {
-    sshLogger.info("Replacing existing base tag", {
-      operation: "opkssh_html_rewrite_base_tag",
-      requestId,
-      basePath,
-    });
-    html = html.replace(/<base[^>]*>/i, baseTag);
-  } else if (html.includes("<head>")) {
-    sshLogger.info("Inserting base tag into head", {
-      operation: "opkssh_html_rewrite_base_tag_insert",
-      requestId,
-      basePath,
-    });
-    html = html.replace(/<head>/i, `<head>${baseTag}`);
-  } else {
-    sshLogger.warn("No <head> tag found, wrapping HTML", {
-      operation: "opkssh_html_rewrite_no_head",
-      requestId,
-      htmlLength: html.length,
-      htmlPreview: html.substring(0, 200),
-    });
-    html = `<!DOCTYPE html><html><head>${baseTag}</head><body>${html}</body></html>`;
-  }
-
-  sshLogger.info("HTML rewrite complete", {
-    operation: "opkssh_html_rewrite_complete",
-    requestId,
-    routePrefix,
-    hasBaseTag: html.includes("<base href="),
-    staticAssetCount: (html.match(/\/static\//g) || []).length,
-  });
-
-  return html;
-}
 
 /**
  * @openapi
